+++ components/QuestionView/index.tsx
@@ -186,6 +186,12 @@ const QuestionView: React.FC<QuestionViewProps> = ({
     setSelectedMultiSelectValues(values);
   };
 
+  const handleSelectChange = (value: string) => {
+    setSelectedSelectValue(value);
+  };
+
+
+  // Handler for date range changes
   const handleDateChange = (
     value: string | DateValue | CalendarDate | null
   ) => {
@@ -283,11 +289,12 @@ const QuestionView: React.FC<QuestionViewProps> = ({
             <SelectboxQuestionComponent
               parsedQuestion={parsed}
               selectedSelectValue={selectedSelectValue}
-              setSelectedSelectValue={setSelectedSelectValue}
+              handleSelectChange={handleSelectChange}
             />
           );
         }
       }
+
       case TEXT_FIELD_QUESTION_TYPE:
         if (parsed.type === 'text') {
 
@@ -503,8 +510,8 @@ const QuestionView: React.FC<QuestionViewProps> = ({
       </ContentContainer>
 
       <SidebarPanel>
-        <p>
-          {trans('bestPractice')}
+        <div className={styles.headerWithLogo}>
+          <h2 className="h4">{Global('bestPractice')}</h2>
           <Image
             className={styles.Logo}
             src="/images/DMP-logo.svg"
@@ -512,7 +519,7 @@ const QuestionView: React.FC<QuestionViewProps> = ({
             height="16"
             alt="DMP Tool"
           />
-        </p>
+        </div>
 
         <ExpandableContentSection
           id="data-description"